+++ packages/nimble-components/src/icon-base/template.ts
@@ -2,7 +2,7 @@ import { html } from '@microsoft/fast-element';
 import type { Icon } from '.';
 
 export const template = html<Icon>`
-    <template slot="start">
+    <template>
         <div class="icon" :innerHTML=${x => x.icon.data}></div>
     </template
 `;