+++ packages/nimble-tokens/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-tokens
 
-This log was last generated on Thu, 13 Jan 2022 19:48:52 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Feb 2022 17:09:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.0-beta.39
+
+Tue, 01 Feb 2022 17:09:06 GMT
+
+### Changes
+
+- Update to latest icons (1458528+fredvisser@users.noreply.github.com)
+
+## 1.0.0-beta.38
+
+Wed, 19 Jan 2022 23:57:51 GMT
+
+### Changes
+
+- Improvements to nimble-tokens README (jattasNI@users.noreply.github.com)
 
 ## 1.0.0-beta.37